First pass at document links. Threaded workspace.
Took a first pass a creating document links for the
`#include` directive. It only works if the file is
relative to the current file and the file exists.

Also created a thread to scan the entire workspace. This
seems to scan pretty quickly, but doesn't really do much
yet. I was not able to get workspace wide diagnostics
to get reported.