Updated parse trees to solve borrow checker.
Updated parse trees to solve borrow checker. After fighting the
BC for over a week, I finally settled on updating the parse trees
to own a copy of the `str` rather than using a reference. The
main issue was I was not able to pass ownership of both the
input text and the parsed lines (using references) together. I
tried to use generics to solve the problem so we can use
references when we don't need to keep the parse tree around for
longer than we have access to the input. However, this resulted
in a lot of the same lifetime issues.

In the end, I decided on using Box<str> because the tree does
not need to mutate the strings and we shouldn't be doing a lot
of cloning. This will increase the amount of memory we use, but
it should still be less than if we use String.

I did add an optino for using thread safe trees, which use
Arc<str>. This should be a drop-in replacement if we every need
that functinoality.